# 8week/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"description": "[목적]\n모든 API의 \"성공 응답\"을 같은 봉투에 담기 위한 헬퍼.\n\n[사용법]\n  컨트롤러에서:\n    return success(user);\n  →  { resultType: \"SUCCESS\", error: null, success: user }  형태로 직렬화됨\n\n[왜 헬퍼 함수인가?]\nTSOA가 응답 직렬화를 담당하기 때문에, 순수 Express처럼 미들웨어 한 줄로\n모든 응답을 wrap할 수가 없음. 컨트롤러 반환 직전에 명시적으로 감싸야 함.\n함수로 만들어두면:\n  1) 형태 변경 시 한 곳만 고치면 됨\n  2) 제네릭 <T> 로 데이터 타입을 보존해서 자동완성도 그대로 동작",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAIL"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {},
							"reason": {
								"type": "string",
								"nullable": true
							},
							"errorCode": {
								"type": "string",
								"nullable": true
							}
						},
						"required": [
							"data",
							"reason",
							"errorCode"
						],
						"type": "object"
					},
					"success": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string",
						"description": "유저 이메일 (로그인 시 사용)"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					},
					"gender": {
						"type": "string",
						"description": "성별 (\"남성\" 또는 \"여성\")"
					},
					"birth": {
						"type": "string",
						"description": "생년월일 (YYYY-MM-DD)"
					},
					"address": {
						"type": "string",
						"description": "기본 주소"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소"
					},
					"phoneNumber": {
						"type": "string",
						"description": "휴대폰 번호"
					},
					"password": {
						"type": "string",
						"description": "비밀번호"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "선호 카테고리 ID 배열 (예: [1, 2])"
					}
				},
				"required": [
					"email",
					"name",
					"gender",
					"birth",
					"address",
					"detailAddress",
					"phoneNumber",
					"password",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MyReviewItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"storeId": {
						"type": "number",
						"format": "double"
					},
					"storeName": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"storeId",
					"storeName",
					"content",
					"rating",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_MyReviewItem_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/MyReviewItem"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PaginatedResponse_MyReviewItem__": {
				"description": "[목적]\n모든 API의 \"성공 응답\"을 같은 봉투에 담기 위한 헬퍼.\n\n[사용법]\n  컨트롤러에서:\n    return success(user);\n  →  { resultType: \"SUCCESS\", error: null, success: user }  형태로 직렬화됨\n\n[왜 헬퍼 함수인가?]\nTSOA가 응답 직렬화를 담당하기 때문에, 순수 Express처럼 미들웨어 한 줄로\n모든 응답을 wrap할 수가 없음. 컨트롤러 반환 직전에 명시적으로 감싸야 함.\n함수로 만들어두면:\n  1) 형태 변경 시 한 곳만 고치면 됨\n  2) 제네릭 <T> 로 데이터 타입을 보존해서 자동완성도 그대로 동작",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/PaginatedResponse_MyReviewItem_"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MyMissionItem": {
				"properties": {
					"userMissionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string",
						"nullable": true
					},
					"startedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"mission": {
						"properties": {
							"deadline": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"reward": {
								"type": "number",
								"format": "double"
							},
							"description": {
								"type": "string",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"deadline",
							"reward",
							"description",
							"name",
							"id"
						],
						"type": "object"
					},
					"store": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"userMissionId",
					"status",
					"startedAt",
					"mission",
					"store"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_MyMissionItem_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/MyMissionItem"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PaginatedResponse_MyMissionItem__": {
				"description": "[목적]\n모든 API의 \"성공 응답\"을 같은 봉투에 담기 위한 헬퍼.\n\n[사용법]\n  컨트롤러에서:\n    return success(user);\n  →  { resultType: \"SUCCESS\", error: null, success: user }  형태로 직렬화됨\n\n[왜 헬퍼 함수인가?]\nTSOA가 응답 직렬화를 담당하기 때문에, 순수 Express처럼 미들웨어 한 줄로\n모든 응답을 wrap할 수가 없음. 컨트롤러 반환 직전에 명시적으로 감싸야 함.\n함수로 만들어두면:\n  1) 형태 변경 시 한 곳만 고치면 됨\n  2) 제네릭 <T> 로 데이터 타입을 보존해서 자동완성도 그대로 동작",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/PaginatedResponse_MyMissionItem_"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CompletedMissionResponse": {
				"properties": {
					"userMissionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string",
						"nullable": true
					},
					"completedMission": {
						"properties": {
							"reward": {
								"type": "number",
								"format": "double"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"reward",
							"name",
							"id"
						],
						"type": "object"
					},
					"store": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"userMissionId",
					"status",
					"completedMission",
					"store"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_CompletedMissionResponse_": {
				"description": "[목적]\n모든 API의 \"성공 응답\"을 같은 봉투에 담기 위한 헬퍼.\n\n[사용법]\n  컨트롤러에서:\n    return success(user);\n  →  { resultType: \"SUCCESS\", error: null, success: user }  형태로 직렬화됨\n\n[왜 헬퍼 함수인가?]\nTSOA가 응답 직렬화를 담당하기 때문에, 순수 Express처럼 미들웨어 한 줄로\n모든 응답을 wrap할 수가 없음. 컨트롤러 반환 직전에 명시적으로 감싸야 함.\n함수로 만들어두면:\n  1) 형태 변경 시 한 곳만 고치면 됨\n  2) 제네릭 <T> 로 데이터 타입을 보존해서 자동완성도 그대로 동작",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"$ref": "#/components/schemas/CompletedMissionResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "8week",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"409": {
						"description": "중복된 이메일 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "회원가입 API",
				"summary": "회원가입을 처리하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "회원가입에 필요한 유저 정보와 선호 카테고리 ID 목록",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest",
								"description": "회원가입에 필요한 유저 정보와 선호 카테고리 ID 목록"
							}
						}
					}
				}
			}
		},
		"/users/{userId}/reviews": {
			"get": {
				"operationId": "HandleListMyReviews",
				"responses": {
					"200": {
						"description": "내 리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedResponse_MyReviewItem__"
								}
							}
						}
					}
				},
				"description": "내가 작성한 리뷰 목록 조회 API",
				"summary": "특정 유저가 작성한 리뷰 목록을 커서 기반으로 조회합니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "리뷰를 조회할 유저 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "마지막으로 조회한 리뷰 ID. 첫 조회 시 생략하거나 0을 전달합니다.",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/missions": {
			"get": {
				"operationId": "HandleListMyOngoingMissions",
				"responses": {
					"200": {
						"description": "내 진행 중 미션 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedResponse_MyMissionItem__"
								}
							}
						}
					}
				},
				"description": "내가 진행 중인 미션 목록 조회 API",
				"summary": "특정 유저가 도전 중인 미션 목록을 커서 기반으로 조회합니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "미션 목록을 조회할 유저 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "마지막으로 조회한 유저 미션 ID. 첫 조회 시 생략하거나 0을 전달합니다.",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/missions/{userMissionId}/complete": {
			"patch": {
				"operationId": "HandleCompleteMyMission",
				"responses": {
					"200": {
						"description": "미션 완료 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_CompletedMissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "진행 중인 미션이 아닌 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "본인의 미션이 아닌 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 미션 도전 기록",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "진행 중인 미션 완료 API",
				"summary": "특정 유저가 도전 중인 미션을 완료 상태로 변경합니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "미션을 완료할 유저 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "완료 처리할 유저 미션 ID",
						"in": "path",
						"name": "userMissionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}